langchain/streamlit projects/requirements.txt
langchain
openai
streamlit
python-dotenv
langchain-openai
faiss-cpu
tiktoken
pypdf
PyPDF2
weaviate-client